Revert "Unlocking of indexes locked in _bt_mergebuild in AccessExclusiveLock"
This reverts commit 181026fd85f1314638ac048649f680b96d9c12f0.

The unlocking of indexes turned out to be unsafe - the buffered writer
is writing the new index under a different relfilenode, but this change
in pg_class is not visible to the other transactions until the commit
due to READ COMMITTED isolation level. This results in 'could not read
block N' kind of errors - it's pretty easy to reproduce.

  1) prepare a large CSV file

     $ for i in `seq 1 10000000`; do echo $i >> /tmp/data.csv; done;

  2) prepare a table with an index

     db=# CREATE TABLE test_tab (id SERIAL PRIMARY KEY);

  3) session A: start the bulkload with BUFFERED/ON_DUPLICATE opts:

     db=# SELECT * FROM pg_bulkload(ARRAY['INPUT=/tmp/data.csv',
                                          'OUTPUT=test_tab',
                                          'LOADER=BUFFERED',
                                          'ON_DUPLICATE_KEEP=NEW',
                                          'DUPLICATE_ERRORS=10000000']);

  4) session B: run a query using the index:

     db=# SET enable_seqscan = off;
     db=# EXPLAIN ANALYZE SELECT * FROM test_tab WHERE (id > 999000);

The first execution finishes OK, the following ones result into errors
when reading the index.